+++ sample/src/main/kotlin/dev/syncforge/sample/navigation/SampleRoutes.kt
@@ -10,3 +10,6 @@ data object NotesRoute
 
 @Serializable
 data object TagsRoute
+
+@Serializable
+data object ConflictSettingsRoute